+++ src/app/presentation/workplace/address/edit-address/address-persona/address-persona.component.ts
@@ -156,8 +156,12 @@ export class AddressPersonaComponent
     this.modalService.resultEvent
       .pipe(takeUntil(this.ngUnsubscribe))
       .subscribe((x: ModalType) => {
-        if (x === ModalType.Delete) {
+        // Only handle delete events from this component context
+        if (x === ModalType.Delete && this.modalService.componentContext === 'address-persona') {
           this.onDeleteCurrentAddress();
+          // Clear context after use
+          this.modalService.componentContext = '';
+          this.modalService.Filing = '';
         }
       });
   }
@@ -470,6 +474,10 @@ export class AddressPersonaComponent
   }
 
   openDeleteAddress() {
+    // Clear modal state and set context to prevent contamination
+    this.modalService.Filing = '';
+    this.modalService.componentContext = 'address-persona';
+    
     this.modalService.deleteMessage = this.message;
     this.modalService.deleteMessageTitle = this.title;
     this.modalService.deleteMessageOkButton = this.title;
